room_wall_visualizer/app/Room_Wall_Visualizer/requirements.txt
# python3 -m pip freeze >> room_wall_visualizer/app/Room_Wall_Visualizer/requirements.txt
affine==2.4.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
clip==0.2.0
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.112.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.23.5
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
langdetect==1.0.9
lightning-utilities==0.11.6
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
onegan==0.5.1
open_clip_torch==2.26.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2024.1
pyvi==0.1.1
PyYAML==6.0.1
rapidfuzz==3.9.4
rasterio==1.3.10
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
six==1.16.0
sklearn-crfsuite==0.5.0
sniffio==1.3.1
snuggs==1.4.7
starlette==0.37.2
sympy==1.13.0
tabulate==0.9.0
tensorboardX==2.6.2.2
Theano==1.0.5
threadpoolctl==3.5.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchmetrics==1.4.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
